# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a the API documentation for Christian Connect a social network site where people can meet and connect.",
    "version": "1.0.0",
    "title": "Christian Connect API",
    "contact": {
      "email": "emesuechinedu@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "christian-connect-api.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/category/video": {
      "get": {
        "tags": [
          "Video"
        ],
        "summary": "Get a list of all video",
        "description": "This endpoint uses a get request to retrieve all video",
        "operationId": "getVideo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Video"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/video/{id}": {
      "get": {
        "tags": [
          "Video"
        ],
        "summary": "Get a video with its comment",
        "description": "This endpoint uses GET request",
        "operationId": "getVideoById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id values that need to be considered for filter",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Video"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/video/:id/comment": {
      "post": {
        "tags": [
          "Video"
        ],
        "summary": "Add a comment to a video",
        "description": "Adds a comment to a video",
        "operationId": "addCommentToVideo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Video_Comment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Video_Comment"
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/lyric": {
      "get": {
        "tags": [
          "Lyric"
        ],
        "summary": "Get a list of all lyrics",
        "description": "This endpoint uses a get request to retrieve all lyrics",
        "operationId": "getLyric",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Lyric"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/lyric/{id}": {
      "get": {
        "tags": [
          "Lyric"
        ],
        "summary": "Get a lyric with its comment",
        "description": "This endpoint uses GET request",
        "operationId": "getLyricById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id values that need to be considered for filter",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Lyric"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/lyric/:id/comment": {
      "post": {
        "tags": [
          "Lyric"
        ],
        "summary": "Add a comment to a lyric",
        "description": "Adds a comment to a lyric",
        "operationId": "addCommentToLyric",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Lyric_Comment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Lyric_Comment"
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
...
    "/category/job": {
      "get": {
        "tags": [
          "Job"
        ],
        "summary": "Get a list of all job",
        "description": "This endpoint uses a get request to retrieve all jobs",
        "operationId": "getJob",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
    "/category/job/{id}": {
      "get": {
        "tags": [
          "Job"
        ],
        "summary": "Get a single",
        "description": "This endpoint uses GET request",
        "operationId": "getAJob",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id values that need to be considered for filter",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          },
          "500": {
            "description": "Something went wrong, please try again later"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Video": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "video_title": {
          "type": "string"
        },
        "video_about": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "video_url": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "uploaded_by": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Video_Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        },
        "comment_by": {
          "type": "string"
        },
        "video_id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Lyric": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "lyric": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "uploaded_by": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Lyric_Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "comment": {
          "type": "string"
        },
        "comment_by": {
          "type": "string"
        },
        "lyric_id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "Job": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "position": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "salary": {
          "type": "string"
        },
        "job_type": {
          "type": "string"
        },
        "deadline": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "requirement": {
          "type": "string"
        },
        "apply": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "uploaded_by": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}